# hra-api-spec.yaml
openapi: 3.0.3
info:
  title: HRA-API
  description: |
    This API provides programmatic access to data registered to the Human Reference Atlas (HRA).
    See the [HuBMAP HRA Portal](https://humanatlas.io/) for details.
  version: 0.18.0
  contact:
    name: HuBMAP Help Desk
    email: help@hubmapconsortium.org
  license:
    name: MIT License
    url: https://spdx.org/licenses/MIT.html
servers:
  - description: HRA-API Production
    url: https://apps.humanatlas.io/api
  - description: HRA-API Staging
    url: https://apps.humanatlas.io/api--staging
  - description: CCF-API (deprecated) Production
    url: https://apps.humanatlas.io/hra-api
  - description: Local Server
    url: /
security: []
...
paths:
...
  /v1/collisions:
    post:
      summary: Given an extraction site, get mesh-based collisions with the reference organ.
      operationId: collisions
      tags:
        - v1
      requestBody:
        $ref: '#/components/requestBodies/ExtractionSite'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
...
  /v1/corridor:
    post:
      summary: Given an extraction site, generate a corridor with the reference organ.
      operationId: corridor
      tags:
        - v1
      requestBody:
        $ref: '#/components/requestBodies/ExtractionSite'
      responses:
        '200':
          description: Successful operation
          content:
            model/gltf-binary:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
...
  /v1/mesh-3d-cell-population:
    post:
      summary: Given a reference organ, 3d scene node, and cell population,  generate cells in that distribution to fit in that 3d scene node.
      operationId: mesh-3d-cell-population
      tags:
        - v1
      requestBody:
        $ref: '#/components/requestBodies/Mesh3dCellPopulation'
      responses:
        '200':
          description: Successful response with CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
...
  /v1/reference-organs:
    get:
      summary: Get all reference organs
      operationId: reference-organs
      tags:
        - v1
      parameters:
        - $ref: '#/components/parameters/Cache'
        - $ref: '#/components/parameters/Token'
      responses:
        '200':
          $ref: '#/components/responses/SpatialEntities'
        '404':
          $ref: '#/components/responses/ErrorMessage'
...
  /v1/sparql:
    get:
      summary: Run a SPARQL query
      operationId: sparql
      tags:
        - v1
      parameters:
        - $ref: '#/components/parameters/Query'
        - $ref: '#/components/parameters/Token'
        - $ref: '#/components/parameters/Format'
      responses:
        '200':
          $ref: '#/components/responses/SparqlResponse'
        '404':
          $ref: '#/components/responses/ErrorMessage'
    post:
      summary: Run a SPARQL query (POST)
      operationId: sparql-post
      tags:
        - v1
      requestBody:
        $ref: '#/components/requestBodies/SparqlQuery'
      parameters:
        - $ref: '#/components/parameters/Token'
        - $ref: '#/components/parameters/Format'
      responses:
        '200':
          $ref: '#/components/responses/SparqlResponse'
        '404':
          $ref: '#/components/responses/ErrorMessage'
...
  /hra-pop/cell-summary-report:
    post:
      summary: Given a cell summary in csv format, retrieve a predicted cell summary report from HRApop showing relative anatomical structures, datasets, and rui locations.
      operationId: cell-summary-report
      tags:
        - hra-pop
      requestBody:
        $ref: '#/components/requestBodies/GetCellSummaryReport'
      responses:
        '200':
          $ref: '#/components/responses/CellSummaryReportResponse'
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
  /hra-pop/rui-location-cell-summary:
    post:
      summary: Given a SpatialEntity already placed relative to a reference SpatialEntity, retrieve a predicted cell summary from HRApop
      operationId: rui-location-cell-summary
      tags:
        - hra-pop
      requestBody:
        $ref: '#/components/requestBodies/GetRuiLocationCellSummary'
      responses:
        '200':
          $ref: '#/components/responses/RuiLocationCellSummaryResponse'
        '400':
          $ref: '#/components/responses/ErrorMessage'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
  /hra-pop/supported-organs:
    get:
      summary: Get all organs supported by HRApop
      operationId: supported-organs
      tags:
        - hra-pop
      responses:
        '200':
          $ref: '#/components/responses/IdLabelsResponse'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
  /hra-pop/supported-reference-organs:
    get:
      summary: Get all reference organs supported by HRApop
      operationId: supported-reference-organs
      tags:
        - hra-pop
      responses:
        '200':
          $ref: '#/components/responses/IdLabelsResponse'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
  /hra-pop/supported-tools:
    get:
      summary: Get all tools supported by HRApop
      operationId: supported-tools
      tags:
        - hra-pop
      responses:
        '200':
          $ref: '#/components/responses/IdLabelsResponse'
        '404':
          $ref: '#/components/responses/ErrorMessage'
        '500':
          $ref: '#/components/responses/ErrorMessage'
...
components:
  schemas:
...
    ErrorMessage:
      title: Error Message
      oneOf:
        - type: string
          title: Error message
          description: Human readable description of the error
        - type: object
          title: ErrorMessageObject
          required:
            - error
          properties:
            error:
              type: string
              description: Human readable description of the error
    JsonLdObject:
      title: JSON-LD Object
      description: Base object for all json-ld objects.
      type: object
      required:
        - '@id'
        - '@type'
      additionalProperties: true
      properties:
        '@id':
          title: Unique Identifier
          type: string
          format: uri
          example: http://purl.obolibrary.org/obo/UBERON_0013702
        '@type':
          title: Object Type
          description: Contains the name of object class/interface
          type: string
          example: Sample
...
    SpatialEntityCommon:
      title: Spatial Entity Common
      description: Properties available in all spatial entity objects.
      type: object
      properties:
        entityId:
          title: Entity Identifier
          description: Identifier of the entity the spatial object represents.
          type: string
          format: uri
        ccf_annotations:
          title: Annotations
          description: Annotations attached to the entity.
          type: array
          items:
            type: string
            format: uri
        representation_of:
          title: Representation
          description: Identifier for the item the spatial object represents.
          type: string
          format: uri
        reference_organ:
          title: Organ
          description: Identifier for the organ the spatial object represents.
          type: string
          format: uri
        sex:
          title: Sex
          type: string
          enum:
            - Female
            - Male
    SpatialEntityCreator:
      title: Spatial Entity Creator
      description: Describes by whom and when an entity object was created.
      type: object
      properties:
        creator:
          title: Full Name
          type: string
        creator_first_name:
          title: First Name
          type: string
        creator_last_name:
          title: Last Name
          type: string
        creator_orcid:
          title: Creator's Identifier
          type: string
        creation_date:
          title: Creation Date
          type: string
          format: date
    SpatialEntityDimensions:
      title: Spatial Entity Dimensions
      description: The dimensions of a spatial entity.
      type: object
      required:
        - x_dimension
        - y_dimension
        - z_dimension
        - dimension_units
      properties:
        x_dimension:
          title: X-Dimension
          type: number
        y_dimension:
          title: Y-Dimension
          type: number
        z_dimension:
          title: Z-Dimension
          type: number
        dimension_units:
          title: Dimension Units
          description: Length units of dimensional values.
          type: string
          default: millimeter
    SpatialPlacementTranslation:
      title: Spatial Placement Translation
      description: The properties to define the Translation aspect of Spatial Placements.
      type: object
      required:
        - x_translation
        - y_translation
        - z_translation
        - translation_units
      properties:
        x_translation:
          title: X-Translation
          type: number
        y_translation:
          title: Y-Translation
          type: number
        z_translation:
          title: Z-Translation
          type: number
        translation_units:
          title: Translation Units
          description: Length units of translation values.
          type: string
          default: millimeter
          enum:
            - millimeter
...
    SpatialPlacementCommon:
      title: Spatial Placement common fields
      description: A Spatial Placement defines how to place a Spatial Entity or Spatial Object Reference relative to another Spatial Entity. It says how to scale, rotate, and translate (in that order) a Spatial Entity or Spatial Object Reference so that it is placed relative to another Spatial Entity.
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonLdObject'
        - $ref: '#/components/schemas/SpatialPlacementTranslation'
        - $ref: '#/components/schemas/SpatialPlacementRotation'
        - $ref: '#/components/schemas/SpatialPlacementScaling'
      required:
        - '@type'
        - target
      properties:
        '@type':
          type: string
          enum:
            - SpatialPlacement
        target:
          type: string
          format: uri
        placement_date:
          title: Date placement was made
          type: string
          format: date
    SpatialEntity:
      title: Spatial Entity
      description: A Spatial Entity defines a bounded cartesian space and the units it is measured in. It typically represents a real world thing, e.g., a human body, a human kidney, a tissue section, or an individual cell.
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonLdObject'
        - $ref: '#/components/schemas/SpatialEntityCommon'
        - $ref: '#/components/schemas/SpatialEntityCreator'
        - $ref: '#/components/schemas/SpatialEntityDimensions'
      required:
        - '@type'
      properties:
        '@type':
          type: string
          enum:
            - SpatialEntity
        label:
          title: Label
          type: string
        comment:
          title: Comment
          type: string
        rui_rank:
          title: RUI Rank
          description: Ranking used to order lists in the RUI.
          type: integer
        sex:
          title: Sex
          type: string
          enum:
            - Female
            - Male
        side:
          title: Side
          description: Used for organs that come in pairs to distinguish between them, i.e. kidneys, etc.
          type: string
          enum:
            - Left
            - Right
        slice_count:
          title: Slice Count
          type: integer
        slice_thickness:
          title: Slice Thickness
          type: number
        placement:
          title: SpatialEntityPlacement
          oneOf:
            - $ref: '#/components/schemas/SpatialPlacement'
            - $ref: '#/components/schemas/SpatialPlacements'
        object:
          $ref: '#/components/schemas/SpatialObjectReference'
        color:
          $ref: '#/components/schemas/Rgba'
    SpatialObjectReference:
      title: Spatial Object Reference
      description: A Spatial Object Reference provides a reference to an external representation of a Spatial Entity.
      type: object
      allOf:
        - $ref: '#/components/schemas/JsonLdObject'
      required:
        - '@type'
      properties:
        '@type':
          type: string
          enum:
            - SpatialObjectReference
        file:
          title: File
          type: string
          format: uri
        file_format:
          title: File Format
          type: string
          example: model/gltf-binary
        file_subpath:
          title: File Subpath
          type: string
    SpatialPlacement:
      title: Spatial Placement
      description: A Spatial Placement defines how to place a Spatial Entity or Spatial Object Reference relative to another Spatial Entity. It says how to scale, rotate, and translate (in that order) a Spatial Entity or Spatial Object Reference so that it is placed relative to another Spatial Entity.
      type: object
      allOf:
        - $ref: '#/components/schemas/SpatialPlacementCommon'
      properties:
        source:
          title: SpacialPlacementSource
          oneOf:
            - $ref: '#/components/schemas/SpatialEntity'
            - $ref: '#/components/schemas/SpatialObjectReference'
        target:
          title: SpacialPlacementTarget
          oneOf:
            - title: Target Sptial Entity IRI
              type: string
              format: uri
              example: http://purl.org/ccf/latest/ccf.owl#VHMLeftKidney
            - $ref: '#/components/schemas/SpatialEntity'
    SpatialPlacements:
      title: Spatial Placements
      type: array
      items:
        $ref: '#/components/schemas/SpatialPlacement'
    Rgba:
      title: RGBA Color Tuple
      type: array
      minItems: 4
      maxItems: 4
      items:
        title: 8-Bit Value
        type: number
        minimum: 0
        maximum: 255
...
    Mesh3dCellPopulationRequest:
      title: mesh-3d-cell-population request body
      type: object
      required:
        - file
        - file_subpath
        - num_nodes
        - node_distribution
      properties:
        file:
          type: string
          format: uri
        file_subpath:
          type: string
        num_nodes:
          type: integer
        node_distribution:
          type: object
          additionalProperties:
            type: number
...
    SparqlQueryRequest:
      title: SPARQL query request body
      type: object
      required:
        - query
      properties:
        query:
          type: string
...
    CellSummaryReportRequest:
      title: Cell summary report request body
      type: object
      required:
        - csvString
      properties:
        csvString:
          title: Csv data
          type: string
        organ:
          title: Organ
          type: string
          format: uri
        tool:
          title: Tool
          type: string
    SourceSimilarityRow:
      title: Source Similarity Row
      type: object
      required:
        - tool
        - modality
        - cell_source
        - cell_source_type
        - cell_source_label
        - cell_source_link
        - similarity
      properties:
        tool:
          title: Tool
          type: string
        modality:
          title: Modality
          type: string
        cell_source:
          title: Cell Source
          type: string
          format: uri
        cell_source_type:
          title: Cell Source Type
          type: string
        cell_source_label:
          title: Cell Source Label
          type: string
        cell_source_link:
          title: Cell Source Link
          type: string
          format: uri
        similarity:
          title: Similarity
          type: number
    CellSummaryReport:
      title: Cell Summary Report
      type: object
      required:
        - sources
        - rui_locations
      properties:
        sources:
          type: array
          items:
            $ref: '#/components/schemas/SourceSimilarityRow'
        rui_locations:
          title: RUI Locations
          type: array
          items: {}
        error:
          title: Error Message
          type: object
          additionalProperties: true
    CellSummaryRow:
      title: Cell Summary Row
      type: object
      required:
        - tool
        - modality
        - cell_id
        - cell_label
        - count
        - percentage
      properties:
        tool:
          title: Tool
          type: string
        modality:
          title: Modality
          type: string
        cell_id:
          title: Cell ID
          type: string
          format: uri
        cell_label:
          title: Cell Label
          type: string
        count:
          title: Count
          type: number
        percentage:
          title: Percentage
          type: number
    IdLabelPair:
      title: ID - Label pair
      type: object
      required:
        - id
        - label
      properties:
        id:
          title: ID (iri) of the entity
          type: string
          format: uri
        label:
          title: Label of the entity
          type: string
...
  parameters:
...
    Cache:
      name: cache
      description: Hint to cache (or not) the response of this request
      in: query
      required: false
      schema:
        type: boolean
        default: true
...
    Token:
      name: token
      description: Authentication token to use for authenticated searches
      in: query
      required: false
      schema:
        type: string
...
    Query:
      name: query
      description: SPARQL query to use
      in: query
      required: true
      schema:
        type: string
        example: SELECT * WHERE { ?sub ?pred ?obj . } LIMIT 10
    Format:
      name: format
      description: Override SPARQL response format (Note that not all formats are supported for all SPARQL query types)
      in: query
      required: false
      schema:
        type: string
        enum:
          - application/json
          - application/ld+json
          - application/n-quads
          - application/n-triples
          - application/sparql-results+json
          - application/sparql-results+xml
          - application/trig
          - simple
          - stats
          - table
          - text/csv
          - text/n3
          - text/tab-separated-values
          - text/turtle
          - tree
...
  responses:
...
    ErrorMessage:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
...
    SpatialEntities:
      description: Successful operation
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SpatialEntity'
...
    SparqlResponse:
      description: Successful operation. SPARQL responses vary by format/content negotiation.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
        text/csv:
          schema:
            type: string
...
    CellSummaryReportResponse:
      description: Successful operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CellSummaryReport'
    RuiLocationCellSummaryResponse:
      description: Successful operation
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CellSummaryRow'
    IdLabelsResponse:
      description: Successful operation
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/IdLabelPair'
...
  requestBodies:
    ExtractionSite:
      description: |
        Request format for routes using posted extraction sites
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpatialEntity'
          example:
            '@context': https://hubmapconsortium.github.io/ccf-ontology/ccf-context.jsonld
            '@id': http://purl.org/ccf/1.5/cc837df1-2788-4510-9c21-9eb0e9b310a5
            '@type': SpatialEntity
            creator: demo demo
            creator_first_name: demo
            creator_last_name: demo
            creation_date: '2024-08-20'
            ccf_annotations:
              - http://purl.obolibrary.org/obo/UBERON_0002015
              - http://purl.obolibrary.org/obo/UBERON_0002189
            x_dimension: 10
            y_dimension: 10
            z_dimension: 10
            dimension_units: millimeter
            placement:
              '@context': https://hubmapconsortium.github.io/ccf-ontology/ccf-context.jsonld
              '@id': http://purl.org/ccf/1.5/cc837df1-2788-4510-9c21-9eb0e9b310a5_placement
              '@type': SpatialPlacement
              target: https://purl.humanatlas.io/ref-organ/kidney-female-right/v1.3#primary
              placement_date: '2024-08-20'
              x_scaling: 1
              y_scaling: 1
              z_scaling: 1
              scaling_units: ratio
              x_rotation: 0
              y_rotation: 0
              z_rotation: 0
              rotation_order: XYZ
              rotation_units: degree
              x_translation: 72.366
              y_translation: 79.067
              z_translation: 34.77
              translation_units: millimeter
...
    Mesh3dCellPopulation:
      description: |
        Request format for requesting a mesh-based 3d cell population.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Mesh3dCellPopulationRequest'
          example:
            file: https://cdn.humanatlas.io/digital-objects/ref-organ/kidney-female-left/v1.3/assets/3d-vh-f-kidney-l.glb
            file_subpath: VH_F_renal_pyramid_L_a
            num_nodes: 10
            node_distribution:
              KEY1: 0.1
              KEY2: 0.3
...
    SparqlQuery:
      description: Request format for a SPARQL query
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SparqlQueryRequest'
          example:
            query: SELECT * WHERE { ?sub ?pred ?obj . } LIMIT 10
        application/x-www-form-urlencoded:
          schema:
            $ref: '#/components/schemas/SparqlQueryRequest'
    GetCellSummaryReport:
      description: Request format for requesting a cell summary report for a given csv file. Must provide a minimum of cell_id and percentage columns.
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CellSummaryReportRequest'
          example:
            organ: http://purl.obolibrary.org/obo/UBERON_0000948
            csvString: |-
              cell_id,cell_label,count,percentage
              http://purl.obolibrary.org/obo/CL_0002131,regular ventricular cardiac myocyte,49273.09543,0.332091136333487
              http://purl.obolibrary.org/obo/CL_0000669,pericyte,24587.941957,0.16571797475566
              http://purl.obolibrary.org/obo/CL_0002144,capillary endothelial cell,20134.675085,0.13570381706955
              http://purl.obolibrary.org/obo/CL_0000057,fibroblast,14931.592835,0.10063604871118
              http://purl.obolibrary.org/obo/CL_0000003,native cell,11003.040295,0.0741583641701745
              http://purl.obolibrary.org/obo/CL_1000413,endothelial cell of artery,6976.051286,0.0470172368606237
              http://purl.obolibrary.org/obo/CL_0000192,smooth muscle cell,4553.526495,0.0306898881601113
              http://purl.obolibrary.org/obo/CL_0000235,macrophage,3870.474525,0.0260862499491858
              http://purl.obolibrary.org/obo/CL_0000115,endothelial cell,2564.473475,0.0172840553851487
              http://purl.obolibrary.org/obo/CL_0002543,vein endothelial cell,1949.697045,0.0131405811128699
              http://purl.obolibrary.org/obo/CL_0000623,natural killer cell,1175.948771,0.00792566734895222
              http://purl.obolibrary.org/obo/CL_0002397,"CD14-positive, CD16-positive monocyte",939.527765,0.00633223548009011
              http://purl.obolibrary.org/obo/CL_0000934,"CD4-positive, alpha-beta cytotoxic T cell",893.222763,0.00602014871959927
              http://purl.obolibrary.org/obo/CL_0002319,neural cell,826.561462,0.00557086443970238
              http://purl.obolibrary.org/obo/CL_0000906,"activated CD8-positive, alpha-beta T cell",603.03561,0.00406434341554534
              http://purl.obolibrary.org/obo/CL_0000794,"CD8-positive, alpha-beta cytotoxic T cell",598.710396,0.00403519230945772
              http://purl.obolibrary.org/obo/CL_1000309,epicardial adipocyte,576.599013,0.003886165863234
              http://purl.obolibrary.org/obo/CL_0001054,CD14-positive monocyte,512.948563,0.00345717413693446
              http://purl.obolibrary.org/obo/CL_0000097,mast cell,507.054682,0.00341745051856579
              http://purl.obolibrary.org/obo/CL_0000576,monocyte,437.848941,0.00295101720503181
              http://purl.obolibrary.org/obo/CL_0000814,mature NK T cell,428.888686,0.00289062681878106
              http://purl.obolibrary.org/obo/CL_0000896,"activated CD4-positive, alpha-beta T cell",314.404462,0.00211902528434063
              http://purl.obolibrary.org/obo/CL_0002138,endothelial cell of lymphatic vessel,253.169304,0.00170631215912871
              http://purl.obolibrary.org/obo/CL_0000236,B cell,239.189975,0.00161209418454692
              http://purl.obolibrary.org/obo/CL_0000451,dendritic cell,200.812358,0.0013534364662104
              http://purl.obolibrary.org/obo/CL_0000077,mesothelial cell,19.717649,0.000132893141887896
    GetRuiLocationCellSummary:
      description: Request format for requesting a cell summary for a given RUI location
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SpatialEntity'
          example:
            '@context': https://hubmapconsortium.github.io/ccf-ontology/ccf-context.jsonld
            '@id': http://purl.org/ccf/1.5/f7130161-ad0a-4801-b5a4-f6297e0f1869
            '@type': SpatialEntity
            creator: Bruce Herr
            creator_first_name: Bruce
            creator_last_name: Herr
            creation_date: '2022-04-27'
            ccf_annotations:
              - http://purl.obolibrary.org/obo/UBERON_0002015
              - http://purl.obolibrary.org/obo/UBERON_0000362
              - http://purl.obolibrary.org/obo/UBERON_0004200
              - http://purl.obolibrary.org/obo/UBERON_0001225
              - http://purl.obolibrary.org/obo/UBERON_0001284
              - http://purl.obolibrary.org/obo/UBERON_0002189
            x_dimension: 11
            y_dimension: 12
            z_dimension: 13
            dimension_units: millimeter
            placement:
              '@context': https://hubmapconsortium.github.io/ccf-ontology/ccf-context.jsonld
              '@id': http://purl.org/ccf/1.5/f7130161-ad0a-4801-b5a4-f6297e0f1869_placement
              '@type': SpatialPlacement
              target: http://purl.org/ccf/latest/ccf.owl#VHFLeftKidneyV1.1
              placement_date: '2022-04-27'
              x_scaling: 1
              y_scaling: 1
              z_scaling: 1
              scaling_units: ratio
              x_rotation: -61
              y_rotation: 50
              z_rotation: -84
              rotation_order: XYZ
              rotation_units: degree
              x_translation: 20.149
              y_translation: 109.963
              z_translation: 38.79
              translation_units: millimeter
...